requirements.txt
flask==2.3.2
flask-cors==3.0.10
requests==2.31.0
numpy==1.24.3
pandas==2.0.2
python-binance==1.0.19
python-dotenv==1.0.0 